rag-granite-qdrant-demo/requirements.txt
qdrant-client
langchain-community
langchain-openai
python-dotenv
requests
langchain
